Extract component CloseIcon from CookieBanner to path ui
